refactor(overlay): rename Dialog to Modal and move it to new overlay/ folder (replaces layout/
