--- frontend/src/components/layout/index.ts
@@ -1,2 +1,0 @@
-export { default as DialogPanel } from './DialogPanel.svelte';
-
