--- frontend/src/components/navbar/popups/MyTrees.svelte
@@ -1,4 +1,0 @@
-<script lang="ts">
-
-</script>
-
