+++ frontend/src/components/overlay/DialogPanel.svelte
