+++ frontend/src/components/overlay/Navbar.svelte
