+++ frontend/src/components/overlay/index.ts
@@ -1,0 +1,4 @@
+export { default as DialogPanel } from './DialogPanel.svelte';
+export { default as Navbar } from './Navbar.svelte';
+export { default as Modal } from './Modal.svelte';
+
