+++ frontend/src/routes/+layout.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
 	import '../app.css';
 	import Map from '../components/map/Map.svelte';
-	import Navbar from '../components/navbar/Navbar.svelte';
+	import { Navbar } from '$components/overlay';
 </script>
 
 <svelte:head>
